Updated the control flow for two types of images
It might be a better practice to handle the formats of different types of images with an assigned category attribute for each image. In this way, you will not need to group the images by their categories manually, and you will not need to change the code when you add new images in the future. My modification is the easiest implementation of the category attribute. It's still not the best practice. Using a separate configuration file to store the information of the images will give you clearer code and easier batch modification of image information.